+++ RELEASE_NOTES.md
@@ -63,6 +63,9 @@ Logtalk equivalents before checking meta-predicate calling compatibility.
 * IMPROVED: Faster generated code for `:/2` non-terminals in the body of
 grammar rules.
 
+* NEW: Support for calling Prolog module meta-predicates that take closures
+as meta-arguments.
+
 Prolog adapter and integration files
 ------------------------------------